Allow stash URLs at subdirectories.
This fix is a stopgap measure and should be reconsidered later.
We want to ensure that stash instances that live at
http://stash.domain/stash
will work. In our current setup only
http://stash.domain/stash
will properly associate the received webhook with the project.
This change fixes it by knowing that the webhook link always has
a `/project/` string immediatley after the end of the stash URL.
This fixes the problem for most instances but if the "project"
string appeared elsewhere, the method would fail.